+++ lippukala/consts.py
@@ -1,0 +1,11 @@
+UNUSED = 0
+USED = 1
+MANUAL_INTERVENTION_REQUIRED = 2
+BEYOND_LOGIC = 3
+
+CODE_STATUS_CHOICES = (
+    (UNUSED, "Unused"),
+    (USED, "Used"),
+    (MANUAL_INTERVENTION_REQUIRED, "Manual intervention required"),
+    (BEYOND_LOGIC, "Beyond logic")
+)
